+++ BuildingReport.API/Controllers/DocumentController.cs
@@ -23,35 +23,58 @@ namespace BuildingReport.API.Controllers
 
         [AllowAnonymous]
         [HttpGet]
-        public List<Document> GetDocuments()
+        public IActionResult GetDocuments()
         {
-            return _documentService.GetAllDocuments();
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            return Ok(_documentService.GetAllDocuments());
         }
 
         [AllowAnonymous]
         [HttpGet("{id}")]
-        public Document GetDocuments(long id)
+        public IActionResult GetDocuments(long id)
         {
-            return _documentService.GetDocumentById(id);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            return Ok(_documentService.GetDocumentById(id));
         }
 
         [AllowAnonymous]
         [HttpGet("building/{buildingId}")]
-        public List<Document> GetDocumentsByBuildingID(long buildingId)
+        public IActionResult GetDocumentsByBuildingID(long buildingId)
         {
-            return _documentService.GetDocumentsByBuildingId(buildingId);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            return Ok(_documentService.GetDocumentsByBuildingId(buildingId));
         }
 
         [AllowAnonymous]
         [HttpGet("user/{userId}")]
-        public List<Document> GetDocumentsByUserID(long userId)
+        public IActionResult GetDocumentsByUserID(long userId)
         {
-            return _documentService.GetDocumentsByUserId(userId);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            return Ok(_documentService.GetDocumentsByUserId(userId));
         }
 
         [HttpPost]
-        public Document Post([FromBody] DocumentDTO documentdto)
+        public IActionResult Post([FromBody] DocumentDTO documentdto)
         {
+            if (documentdto == null)
+            {
+                return BadRequest(ModelState);
+            }
+
+            if (_documentService.DocumentExists(documentdto.Report))
+            {
+                ModelState.AddModelError("", "Document already exists");
+                return StatusCode(422, ModelState);
+            }
+
             var document = new Document()
             {
                 Id = documentdto.Id,
@@ -63,11 +86,11 @@ namespace BuildingReport.API.Controllers
 
             };
 
-            return _documentService.CreateDocument(document);
+            return Ok(_documentService.CreateDocument(document));
         }
 
         [HttpPut]
-        public Document Put([FromBody] DocumentDTO documentdto)
+        public IActionResult Put([FromBody] DocumentDTO documentdto)
         {
             var document = new Document()
             {
@@ -80,13 +103,21 @@ namespace BuildingReport.API.Controllers
 
             };
 
-            return _documentService.UpdateDocument(document);
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            return Ok(_documentService.UpdateDocument(document));
         }
 
         [HttpDelete("{id}")]
-        public void Delete(long id)
+        public IActionResult Delete(long id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             _documentService.DeleteDocument(id);
+
+            return NoContent();
         }
     }
 }
